First (and presumably only) commit
This is the source code of the homework
assignment that we made during our class.

It shows how to make forms in rails (using the
form_tag) that include the authenticity_token.

It also contains an example of a blog and blogs controller
that also shows how you can use a form_tag for
a specific rails model.